[add] api 타입 common 폴더 이동
