+++ src/api/reservation/common.ts
@@ -30,3 +30,36 @@ export interface ReservationResponse {
     endSchedule: string;
   }[];
 }
+
+export interface ReservationListParams {
+  'reservation-status'?: '대기' | '취소' | '승인' | '거절';
+  page?: number;
+  size?: number;
+  sort?: string;
+}
+export interface GetRequestInformationParams {
+  storeId: string;
+}
+
+export interface SeatScheduleParams {
+  storeChairId: number;
+  startSchedule: string;
+  endSchedule: string;
+  customUtilizationContents: {
+    fieldId: number;
+    content: string[];
+  }[];
+}
+export interface SpaceScheduleParams {
+  storeSpaceId: number;
+  startSchedule: string;
+  endSchedule: string;
+  customUtilizationContents: {
+    fieldId: number;
+    content: string[];
+  }[];
+}
+
+export interface ReservationParams {
+  'reservation-date-and-time': string;
+}
